moving flask applicatin to its own folder
