+++ log_analyzer/src/gui/application.py
