+++ log_analyzer/src/gui/static/css/main.css
