+++ log_analyzer/src/gui/static/js/channels.js
