+++ log_analyzer/src/gui/static/js/channels_map.js
