+++ log_analyzer/src/gui/static/js/charts.js
