+++ log_analyzer/src/gui/static/js/index.js
