+++ log_analyzer/src/gui/static/js/measurements.js
