+++ log_analyzer/src/gui/static/js/utils.js
