+++ log_analyzer/src/gui/templates/base.html
