+++ log_analyzer/src/gui/templates/index.html
