+++ log_analyzer/src/gui/templates/measurements.html
