+++ log_analyzer/src/gui/templates/settings.html
